requirements.txt
anytree==2.12.1
astrapy>=1.0.0
cassio>=0.1.10
coloredlogs==15.0.1
google-cloud-storage>=2.10.0
pypdf>=4.0.0
pymupdf>=1.23.0
dash==2.18.1
dash_bootstrap_components==1.6.0
dash-loading-spinners==1.0.3
gliner==0.2.13
html2text==2024.2.26
keybert==0.8.5
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai==0.2.10
langchain-text-splitters>=0.3.0,<0.4.0
openai>=1.54.0
python-dotenv==1.0.1
ragstack-ai-knowledge-store==0.2.1
tabulate==0.9.0
unstructured==0.15.13
